blob: use uint64_t for unmap and write_zeroes lba count
Previous patches (5363eb3c) tried to work around the
32-bit unmap and write_zeroes LBA counts by breaking
up larger operations into smaller chunks of max size
UINT32_MAX lba chunks.

But some SSDs may just ignore unmap operations that
are not aligned to full physical block boundaries -
and a UINT32_MAX lba unmap on a 512B logical /
4KiB physical SSD would not be aligned.  If the SSD
decided to ignore the unmap/deallocate (which it is
allowed to do according to NVMe spec), we could end
up with not unmapping *any* blocks.  Probably SSDs
should always be trying hard to unmap as many
blocks as possible, but let's not try to depend on
that in blobstore.

So one option would be to break them into chunks
close to UINT32_MAX which are still aligned to
4KiB boundaries.  But the better fix is to just
change the unmap and write_zeroes APIs to take
64-bit arguments, and then we can avoid the
chunking altogether.

Fixes issue #2190.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/9737
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ include/spdk/blob.h
@@ -184,11 +184,11 @@ struct spdk_bs_dev {
 		      struct spdk_bs_dev_cb_args *cb_args);
 
 	void (*write_zeroes)(struct spdk_bs_dev *dev, struct spdk_io_channel *channel,
-			     uint64_t lba, uint32_t lba_count,
+			     uint64_t lba, uint64_t lba_count,
 			     struct spdk_bs_dev_cb_args *cb_args);
 
 	void (*unmap)(struct spdk_bs_dev *dev, struct spdk_io_channel *channel,
-		      uint64_t lba, uint32_t lba_count,
+		      uint64_t lba, uint64_t lba_count,
 		      struct spdk_bs_dev_cb_args *cb_args);
 
 	struct spdk_bdev *(*get_base_bdev)(struct spdk_bs_dev *dev);
