+++ lib/blob/zeroes.c
@@ -85,7 +85,7 @@ zeroes_writev(struct spdk_bs_dev *dev, struct spdk_io_channel *channel,
 
 static void
 zeroes_write_zeroes(struct spdk_bs_dev *dev, struct spdk_io_channel *channel,
-		    uint64_t lba, uint32_t lba_count,
+		    uint64_t lba, uint64_t lba_count,
 		    struct spdk_bs_dev_cb_args *cb_args)
 {
 	cb_args->cb_fn(cb_args->channel, cb_args->cb_arg, -EPERM);
@@ -94,7 +94,7 @@ zeroes_write_zeroes(struct spdk_bs_dev *dev, struct spdk_io_channel *channel,
 
 static void
 zeroes_unmap(struct spdk_bs_dev *dev, struct spdk_io_channel *channel,
-	     uint64_t lba, uint32_t lba_count,
+	     uint64_t lba, uint64_t lba_count,
 	     struct spdk_bs_dev_cb_args *cb_args)
 {
 	cb_args->cb_fn(cb_args->channel, cb_args->cb_arg, -EPERM);
